First commit with code: complete decoder
This is a C rewrite of an incomplete C++ NES emulator. I opted for this
rewrite just for fun, honestly; that's the whole point of personal
projects. My first design decision that differs from the old version is
precisely that of separating the decoder from the rest of the processor.
I reckon this is way cleaner and will facilitate such fancy things as a
disassembler.