Add simple disassembler, for testing and profit
Not testing the program until this very moment has clearly been a bit of
a bad idea. It seems I've messed up decoding somehow. Admitedlly, it is a
quite complicated logic, which is to say, quite easy to mess up. I'll
revise it all as soon as I have some free time.