Merge the registers back into the main PPU module
So, I read a bit more about software design and yes, that was a bad
idea. I've also hidden the information that there are only eight
registers (for now) within the PPU. As it standed, that information was
leaking into the main bus implementation. I'm sure there is more I could
improve, but that's what I could think of today.