Pivot to being just a 6502 simulation
I haven't touched this project in a while, but I would like to go back
to working on it occasionally. While thinking about the other day, I
realized that I had a blast writing stuff for the processor simulation,
but the PPU was a pain in the ass. I just don't have much fun reading
about PPU registers and sprite specs and how to convert all that to SDL
and whatnot.

So, in the interest of my enjoyment and interest, and inspired by the
recent Ben Eater videos on the 6502, I will be focusing just on making
the processor stuff the best I can. I'll make this flexibe enough to use
in all kinds of emulators for 6502 computers.